Undo my error handling so we can do some hack in the proxy
It used to call auth_ok() with no data and then crash in auth_ok()
My old error handling commit made it crash before and never call auth_ok()
Now this commit is reverted again so that I can just overwrite auth_ok()
and make it not crash :D

The reason it crashed is because it tried to get the uuid from the empty data.
We can just check if the data is empty and then use the uuid from the credentials the user provided.